some reorg and adding the builtin conductor class.
