+++ src/canary_hpc/executor.py
@@ -1,18 +1,24 @@
 # Copyright NTESS. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: MIT
-
+import argparse
 import json
 import os
+import threading
 from typing import Any
+from typing import Sequence
 
 import hpc_connect
 
 import canary
+from _canary.plugins.builtin.conductor import Runner
 from _canary.plugins.types import Result
+from _canary.queue import ResourceQueue
+from _canary.queue import process_queue
 
 from .testbatch import TestBatch
 
+global_lock = threading.Lock()
 logger = canary.get_logger(__name__)
 
 
@@ -31,7 +37,7 @@ class CanaryHPCExecutor:
             cases = TestBatch.loadindex(self.batch)
             self.cases.extend(cases)
 
-    def setup(self, *, config: canary.Config) -> None:
+    def setup(self, config: canary.Config) -> None:
         pool = self.generate_resource_pool()
         config.resource_pool.fill(pool)
         stage = TestBatch.stage(self.batch)
@@ -43,6 +49,9 @@ class CanaryHPCExecutor:
         if not os.path.exists(f):
             with open(f, "w") as fh:
                 self.backend.config.dump(fh)
+
+    def register(self, pluginmanager: canary.CanaryPluginManager) -> None:
+        pluginmanager.register(self, "canary_hpc_executor")
 
     def generate_resource_pool(self) -> dict[str, Any]:
         # set the resource pool for this backend
@@ -60,9 +69,16 @@ class CanaryHPCExecutor:
         }
         return pool
 
-    @property
-    def case_specs(self) -> list[str]:
-        return [f"/{case}" for case in self.cases]
+    def run(self, args: argparse.Namespace) -> int:
+        n = len(self.cases)
+        logger.info(f"Selected {n} {canary.string.pluralize('test', n)} from batch {args.batch_id}")
+        case_specs = [f"/{case}" for case in self.cases]
+        session = canary.Session.casespecs_view(os.getcwd(), case_specs)
+        session.run()
+        canary.config.pluginmanager.hook.canary_runtests_summary(
+            cases=session.active_cases(), include_pass=False, truncate=10
+        )
+        return session.exitstatus
 
     @canary.hookimpl
     def canary_resource_count(self, type: str) -> int:
@@ -74,10 +90,34 @@ class CanaryHPCExecutor:
 
     @canary.hookimpl
     def canary_resources_avail(self, case: canary.TestCase) -> Result:
-        # The resource pool was already set above, so we can just leverage it
         return canary.config.resource_pool.accommodates(case)
 
     @canary.hookimpl
     def canary_resource_types(self) -> list[str]:
-        # The resource pool was already set above, so we can just leverage it
         return canary.config.resource_pool.types
+
+    @canary.hookimpl
+    def canary_runtests(self, cases: Sequence["canary.TestCase"]) -> int:
+        """Run each test case in ``cases``.
+
+        Args:
+        cases: test cases to run
+
+        Returns:
+        The session returncode (0 for success)
+
+        """
+        queue = ResourceQueue.factory(global_lock, cases, resource_pool=canary.config.resource_pool)
+        runner = Runner()
+        return process_queue(queue, runner)
+
+    @staticmethod
+    def setup_parser(parser: canary.Parser) -> None:
+        parser.add_argument(
+            "--workers", type=int, help="Run tests in batch using this many workers"
+        )
+        parser.add_argument(
+            "--backend", dest="canary_hpc_backend", help="The HPC connect backend name"
+        )
+        parser.add_argument("--case", dest="canary_hpc_case", help="Run only this case")
+        parser.add_argument("batch_id")
